fix: rename CmcdParam file to match exported const
