+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.5] - 2023-10-12
 
+### Fixed
+- `CMCD_PARAM` constant does not match file name [#27](https://github.com/streaming-video-technology-alliance/common-media-library/issues/27)
+
 
 ## [0.4.4] - 2023-10-11
