+++ lib/src/cmcd/CMCD_PARAM.ts
